* Updated the NativeVideoView to have delegated functionality
